clairvoyant policy
