+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/allegro/bigcache v1.2.1-0.20190218064605-e24eb225f156
 	github.com/coocood/freecache v1.1.0
 	github.com/davecgh/go-spew v1.1.1
+	github.com/emirpasic/gods v1.12.0
 	github.com/goburrow/cache v0.1.0
 	github.com/golang/groupcache v0.0.0-20190129154638-5b532d6fd5ef
 	github.com/stretchr/testify v1.3.0
